@@ -29,7 +29,7 @@ namespace SimplCommerce.Module.Core
             ConfigurePolicies(services);
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
         }
 
