@@ -10,7 +10,7 @@ namespace SimplCommerce.Module.StorageAzureBlob
 {
     public class ModuleInitializer : IModuleInitializer
     {
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
         }
 
