@@ -8,12 +8,13 @@ using SimplCommerce.Module.Core.Extensions;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using SimplCommerce.Infrastructure;
+using Microsoft.Extensions.Hosting;
 
 namespace SimplCommerce.WebHost.Extensions
 {
     public static class ApplicationBuilderExtensions
     {
-        public static IApplicationBuilder RunModuleConfigures(this IApplicationBuilder app, IHostingEnvironment env)
+        public static IApplicationBuilder RunModuleConfigures(this IApplicationBuilder app, IWebHostEnvironment env)
         {
             var moduleInitializers = app.ApplicationServices.GetServices<IModuleInitializer>();
             foreach (var moduleInitializer in moduleInitializers)
@@ -26,6 +27,7 @@ namespace SimplCommerce.WebHost.Extensions
         public static IApplicationBuilder UseCustomizedIdentity(this IApplicationBuilder app)
         {
             app.UseAuthentication();
+            app.UseAuthorization();
             return app;
         }
 
@@ -42,7 +44,7 @@ namespace SimplCommerce.WebHost.Extensions
             return app;
         }
 
-        public static IApplicationBuilder UseCustomizedStaticFiles(this IApplicationBuilder app, IHostingEnvironment env)
+        public static IApplicationBuilder UseCustomizedStaticFiles(this IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
