@@ -1,7 +1,5 @@
 ﻿@model LoginViewModel
 @inject SignInManager<User> SignInManager
-@using Microsoft.AspNetCore.Authentication
-@using Microsoft.AspNetCore.Authentication.OpenIdConnect
 
 @{
     ViewData["Title"] = Localizer["Log in"];
@@ -53,7 +51,7 @@
             </form>
         </section>
     </div>
-    <div class="col-md-4">
+    @*<div class="col-md-4">
         <section>
             <h4>@Localizer["Use another service to log in."]</h4>
             <hr />
@@ -84,7 +82,7 @@
                 }
             }
         </section>
-    </div>
+    </div>*@
 </div>
 
 @section Scripts {
