@@ -1,9 +1,11 @@
 ﻿using System;
 using System.IO;
 using System.Reflection;
+using Autofac.Extensions.DependencyInjection;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Serilog;
 using SimplCommerce.Module.Core.Extensions;
@@ -14,24 +16,22 @@ namespace SimplCommerce.WebHost
     public class Program
     {
         public static void Main(string[] args) =>
-            CreateWebHostBuilder(args)
+            CreateHostBuilder(args)
                 .ConfigureLogging(SetupLogging)
                 .UseSerilog()
-                .UseApplicationInsights()
                 .Build()
                 .Run();
 
         // For EF to instantiate DbContext object. "BuildWebHost" is a convention!
-        private static IWebHost BuildWebHost(string[] args) =>
-            CreateWebHostBuilder(args).Build();
+        private static IHost BuildWebHost(string[] args) => CreateHostBuilder(args).Build();
 
-        private static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            Microsoft.AspNetCore.WebHost
-                .CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
+        private static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .UseServiceProviderFactory(new AutofacServiceProviderFactory())
+                .ConfigureWebHostDefaults(webBuilder => webBuilder.UseStartup<Startup>())
                 .ConfigureAppConfiguration(SetupConfiguration);
 
-        private static void SetupLogging(WebHostBuilderContext context, ILoggingBuilder loggingBuilder)
+        private static void SetupLogging(HostBuilderContext context, ILoggingBuilder loggingBuilder)
         {
             var configuration = new ConfigurationBuilder()
                 .SetBasePath(Directory.GetCurrentDirectory())
@@ -44,7 +44,7 @@ namespace SimplCommerce.WebHost
                 .CreateLogger();
         }
 
-        private static void SetupConfiguration(WebHostBuilderContext context, IConfigurationBuilder configBuilder)
+        private static void SetupConfiguration(HostBuilderContext context, IConfigurationBuilder configBuilder)
         {
             configBuilder
                 .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
