@@ -10,7 +10,7 @@ namespace SimplCommerce.Test.Shared.MockQueryable
         {
             var mock = new Mock<IQueryable<TEntity>>();
             var enumerable = new TestAsyncEnumerable<TEntity>(data);
-            mock.As<IAsyncEnumerable<TEntity>>().Setup(d => d.GetEnumerator()).Returns(enumerable.GetEnumerator);
+            mock.As<IAsyncEnumerable<TEntity>>().Setup(d => d.GetAsyncEnumerator(default)).Returns(enumerable.GetEnumerator);
             mock.As<IQueryable<TEntity>>().Setup(m => m.Provider).Returns(enumerable);
             mock.As<IQueryable<TEntity>>().Setup(m => m.Expression).Returns(data?.Expression);
             mock.As<IQueryable<TEntity>>().Setup(m => m.ElementType).Returns(data?.ElementType);
